--- test/crashThisApp.test.js
@@ -1,7 +1,0 @@
-import crashThisApp from '../lib/crashThisApp';
-
-describe('crash function', () => {
-  it('it works correctly', () => {
-    crashThisApp();
-  });
-});
